chore: clean up feedback state machine and move to own file
